lod2/src/requirements.txt
addict==2.4.0
affine==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.5.0
ansicon==1.89.0
anyio==3.7.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
croniter==1.4.1
cryptography==41.0.4
cycler==0.12.0
dateutils==0.6.12
deepdiff==6.5.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
Fiona==1.9.4.post1
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
geopandas==0.13.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
jmespath==0.10.0
kiwisolver==1.4.5
lightning==2.0.6
lightning-cloud==0.5.39
lightning-utilities==0.9.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mmcv==2.0.1
mmdet==3.1.0
mmengine==0.8.5
mmsegmentation==1.1.1
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.0
opencv-contrib-python==4.8.0.74
opencv-python==4.7.0.72
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.26
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.0
platformdirs==3.11.0
prettytable==3.9.0
psutil==5.9.5
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.0
pyshp==2.3.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
rasterio==1.3.9
readchar==4.0.5
regex==2023.10.3
requests==2.28.2
rich==13.4.2
scipy==1.11.3
shapely==2.0.1
shapelysmooth==0.1.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
tomli==2.0.1
torchmetrics==1.2.0
tqdm==4.65.0
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
uvicorn==0.23.2
wcwidth==0.2.8
websocket-client==1.6.3
websockets==11.0.3
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
